refactored_web_app/requirements.txt
flask
numpy
opencv-python
insightface
pillow
scikit-learn
pydantic
